+++ examples/bank-account/src/it/resources/application.conf
@@ -1,0 +1,6 @@
+es4kafka {
+    service {
+        applicationId = "bank-it"
+    }
+    kafkaStreams.cleanUpState = true
+}
